requirements.txt
fastapi
uvicorn
requests
aiohttp
python-dotenv
pytest
pytest-asyncio
httpx